+++ src/prefect/schedules/__init__.py
@@ -2,4 +2,10 @@ import prefect.schedules.clocks
 import prefect.schedules.filters
 import prefect.schedules.adjustments
 import prefect.schedules.schedules
-from prefect.schedules.schedules import Schedule, IntervalSchedule, CronSchedule
+from prefect.schedules.schedules import (
+    Schedule,
+    IntervalSchedule,
+    CronSchedule,
+    OneTimeSchedule,
+    UnionSchedule,
+)